Cleaned up project, fixed api calls, and updating logic
