+++ src/config/urls.js
@@ -1,11 +1,12 @@
-export const urls = {
+export const localUrls = {
   profileData: "http://localhost:8080/info",
   uploadImage: "http://localhost:8080/info/upload",
   uploadImageB64: "http://localhost:8080/info/userUploadedImagesB64",
   updateBasicInfo: "http://localhost:8080/info/updateAllBasicInfo",
+  deleteImage: "http://localhost:8080/info/removeImage",
 };
 
-export const apiUrls = {
+export const urls = {
   profileData: "https://afternoon-lake-11970-27bd8a5aced0.herokuapp.com/info",
   uploadImage:
     "https://afternoon-lake-11970-27bd8a5aced0.herokuapp.com/info/upload",
@@ -13,4 +14,6 @@ export const apiUrls = {
     "https://afternoon-lake-11970-27bd8a5aced0.herokuapp.com/info/userUploadedImagesB64",
   updateBasicInfo:
     "https://afternoon-lake-11970-27bd8a5aced0.herokuapp.com/info/updateAllBasicInfo",
+  deleteImage:
+    "https://afternoon-lake-11970-27bd8a5aced0.herokuapp.com/info/removeImage",
 };
