--- src/styling/likes.scss
@@ -1,41 +1,0 @@
-$text-color: white;
-$background: #242526;
-
-.likes{
-    width: 100%;
-    height: 100vh;
-    background-color: $background;
-
-    &-header{
-        color: $text-color;
-
-        h1{
-            text-align: center;
-        }
-    }
-
-    &-nav{
-        position: fixed;
-        bottom: 0;
-        width: 100%;
-    }
-}
-
-.activity{
-    flex-flow: row nowrap;
-    align-items: center;
-    text-overflow: ellipsis;
-    div{
-        padding: 5px 7px;
-    }
-    img{
-        width:50px;
-        height: 50px;
-        border-radius: 50%;
-    }
-
-    h2{
-        color:white;
-        font-size: 15px;
-    }
-}
